Cleanups from stricter span usage of size_t, pt. 2: net
More use of first(), compile-time sizes, size_t, etc.

Bug: none
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6047031
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1387584}

+++ net/base/upload_bytes_element_reader.h
@@ -6,6 +6,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #ifndef NET_BASE_UPLOAD_BYTES_ELEMENT_READER_H_
 #define NET_BASE_UPLOAD_BYTES_ELEMENT_READER_H_
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include <string>
@@ -42,7 +43,7 @@ class NET_EXPORT UploadBytesElementReader : public UploadElementReader {
 
  private:
   const base::raw_span<const uint8_t, DanglingUntriaged> bytes_;
-  uint64_t offset_ = 0;
+  size_t offset_ = 0;
 };
 
 // A subclass of UplodBytesElementReader which owns the data given as a vector.
